pyproject.toml
[project]
name = "ai-engineer-club-nc"
version = "0.1.0"
description = "AI Engineer Club Assignment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.98.0",
    "openai-agents>=0.10.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
    "google-adk>=1.12.0",
    "google-genai>=1.31.0",
    "litellm>=1.76.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
]
